+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/any.kt
@@ -1,7 +1,11 @@
 package io.kotest.assertions
 
+import io.kotest.common.ExperimentalKotest
+
 /**
  * Runs multiple assertions and expects at least one to succeed, will suppress all exceptions otherwise.
+ *
+ * TODO: add experimental annotation
  *
  * ```
  *   any {
@@ -11,19 +15,21 @@ package io.kotest.assertions
  *   }
  * ```
  */
+@ExperimentalKotest
 suspend inline fun <T> any(crossinline assertions: suspend () -> T): T? {
    val (result, failures, assertionCount) = errorScope { assertions() }
+   assertionCounter.inc(assertionCount)
 
-   if (failures.isEmpty() && assertionCount == 0 || assertionCount > failures.size) {
+   if (assertionCount > failures.size || failures.isEmpty() && assertionCount == 0) {
       return result.getOrNull()
    }
 
-   assertionCounter.set(assertionCount)
-   errorCollector.collectOrThrow(failures + failure("Any expected at least one assertion to succeed but they all failed"))
-
+   errorCollector.pushErrors(failures)
+   errorCollector.collectOrThrow(failure("Any expected at least one assertion to succeed but they all failed"))
    return null
 }
 
+@ExperimentalKotest
 suspend inline fun <T> any(t: T, crossinline assertions: suspend T.(T) -> Unit) = any {
    t.assertions(t)
    t