+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/either.kt
@@ -1,7 +1,11 @@
 package io.kotest.assertions
 
+import io.kotest.common.ExperimentalKotest
+
 /**
  * Runs multiple assertions and expects exactly one to succeed, will suppress all exceptions otherwise.
+ *
+ * TODO: add experimental annotation
  *
  * ```
  *   either {
@@ -11,31 +15,36 @@ package io.kotest.assertions
  *   }
  * ```
  */
+@ExperimentalKotest
 suspend inline fun <T> either(crossinline assertions: suspend () -> T): T? {
-   val (result, failures, assertionsCounter) = errorScope { assertions() }
+   val (result, failures, assertionsCount) = errorScope { assertions() }
+   assertionCounter.inc(assertionsCount)
 
-   if (assertionsCounter < 2) {
+   if (assertionsCount < 2) {
       errorCollector.collectOrThrow(failures + failure("Either cannot ensure a mutual exclusion with less than two assertions"))
    }
 
-   if (assertionsCounter == failures.size + 1) {
+   if (assertionsCount == failures.size + 1) {
       return result.getOrNull()
    }
 
-   assertionCounter.set(assertionsCounter)
+   errorCollector.pushErrors(failures)
 
-   when {
-      assertionsCounter > failures.size + 1 ->
-         errorCollector.collectOrThrow(failures + failure("Either expected a single assertion to succeed but more than one succeeded"))
-      assertionsCounter < failures.size - 1 ->
-         errorCollector.collectOrThrow(failures + failure("Either expected a single assertion to succeed but they all failed"))
-      else ->
-         errorCollector.collectOrThrow(failures)
+   val f = when {
+      assertionsCount == failures.size -> failure("Either expected a single assertion to succeed, but none succeeded.")
+      assertionsCount > failures.size + 1 -> failure("Either expected a single assertion to succeed, but more than one succeeded.")
+      else -> failure("Either expected a single assertion to succeed, but there were more failures than assertions.")
    }
 
+   errorCollector.collectOrThrow(f)
    return null
 }
 
+/**
+ * Runs multiple assertions and expects exactly one to succeed, will suppress all exceptions otherwise.
+ * Returns the original value [t] on success for use in subsequent assertions.
+ */
+@ExperimentalKotest
 suspend inline fun <T> either(t: T, crossinline assertions: suspend T.(T) -> Unit) = either {
    t.assertions(t)
    t